Maven migration and rename packages
