+++ .gitignore
@@ -1,9 +1,3 @@
-.gradle
-build/
-!gradle/wrapper/gradle-wrapper.jar
-!**/src/main/**/build/
-!**/src/test/**/build/
-
 ### IntelliJ IDEA ###
 .idea/
 *.iws
@@ -37,3 +31,5 @@ bin/
 
 ### Mac OS ###
 .DS_Store
+
+target/
