--- build.gradle.kts
@@ -1,19 +1,0 @@
-plugins {
-    id("java")
-}
-
-group = "br.com.tony.normalizer"
-version = "1.0-SNAPSHOT"
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation(platform("org.junit:junit-bom:5.10.0"))
-    testImplementation("org.junit.jupiter:junit-jupiter")
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
