--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Wed Jul 31 18:41:46 BRT 2024
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.5-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
