--- settings.gradle.kts
@@ -1,2 +1,0 @@
-rootProject.name = "data-normalizer"
-
