+++ src/main/java/br/com/normalizer/processor/NormalizeProcessorForValue.java
@@ -1,4 +1,4 @@
-package br.com.tony.normalizer.processor;
+package br.com.normalizer.processor;
 
 
 public interface NormalizeProcessorForValue<T> {
