+++ src/main/java/br/com/normalizer/processor/Result.java
@@ -1,0 +1,4 @@
+package br.com.normalizer.processor;
+
+
+public record Result<T>(T value) { }
