+++ src/main/java/br/com/normalizer/rules/Rule.java
@@ -1,0 +1,8 @@
+package br.com.normalizer.rules;
+
+
+import br.com.normalizer.processor.Result;
+
+public interface Rule<T> {
+    Result<T> apply(T value);
+}
