--- src/main/java/br/com/tony/normalizer/processor/Result.java
@@ -1,4 +1,0 @@
-package br.com.tony.normalizer.processor;
-
-
-public record Result<T>(T value) { }
