--- src/main/java/br/com/tony/normalizer/rules/Rule.java
@@ -1,8 +1,0 @@
-package br.com.tony.normalizer.rules;
-
-
-import br.com.tony.normalizer.processor.Result;
-
-public interface Rule<T> {
-    Result<T> apply(T value);
-}
